Update copyright and extract interfaces
