+++ static/alert.ts
@@ -1,4 +1,4 @@
-// Copyright (c) 2017, Compiler Explorer Authors
+// Copyright (c) 2021, Compiler Explorer Authors
 // All rights reserved.
 //
 // Redistribution and use in source and binary forms, with or without
@@ -24,51 +24,7 @@
 
 import $ from 'jquery';
 
-export interface AlertAskOptions {
-    /** Function to execute on "yes" button press */
-    yes?: () => void;
-    /** Function to execute on "no" button press */
-    no?: () => void;
-    /** HTML markup of "yes" button */
-    yesHtml?: string;
-    /** Custom HTML class to add to "yes" button */
-    yesClass?: string;
-    /** HTML markup of "no" button */
-    noHtml?: string;
-    /** Custom HTML class to add to "no" button */
-    noClass?: string;
-    /** Function to execute on pane closure */
-    onClose?: () => void;
-}
-
-export interface AlertNotifyOptions {
-    /**
-     * Which group this notification is from. Sets data-group attribute value
-     * Default: ""
-     */
-    group?: string
-    /** If set to true, other notifications within the same group will be removed before sending this one. (Note that
-     * this only has any effect if options.group is set).
-     * Default: true
-     */
-    collapseSimilar?: boolean
-    /**
-     * Space separated list of HTML classes to give to the notification's div element.
-     * Default: ""
-     */
-    alertClass?: string
-    /**
-     * If set to true, the notification will fade out and be removed automatically.
-     * Default: true
-     */
-    autoDismiss?: boolean
-    /**
-     * If allow by autoDismiss, controls how long the notification will be visible (in milliseconds) before
-     * automatically removed
-     * Default: 5000
-     */
-    dismissTime?: number
-}
+import { AlertAskOptions, AlertNotifyOptions } from "./alert.interfaces";
 
 export class Alert {
     yesHandler: () => void | null = null;
